restructure files to prepare for dashboard
