+++ src/App.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import Navigation from './components/Navigation';
+
 import './App.css';
 
 class App extends Component {
