--- src/components/Dashboard.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Dashboard extends Component {
-    render() {
-        return (
-            <div className="dashboard">
-                Customer Dashboard
-            </div>
-        );
-    }
-}
-
-export default Dashboard;
