+++ src/components/Navigation.js
@@ -1,5 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import { Link } from 'react-router';
+
 import './Navigation.css';
 
 class Navigation extends Component {
@@ -10,18 +11,12 @@ class Navigation extends Component {
                 <div className="flex-grow pa3 flex items-center">
                     <Link to="/" className="f6 link dib dim mr3 mr4-ns" title="Home">Home</Link>
                     <Link to="/about" className="f6 link dib dim mr3 mr4-ns" title="About">About</Link>
-                    <Link to="/bloodlines" className="f6 link dib dim mr3 mr4-ns" title="Bloodlines">Bloodlines</Link>
+                    <Link to="/dashboard" className="f6 link dib dim mr3 mr4-ns" title="Dashboard">Dashboard</Link>
                     <Link to="/login" className="f6 link dib dim mr3 mr4-ns" title="Login">Login</Link>
                 </div>
             </nav>
         );
     }
-}
-
-Navigation.propTypes = {
-    isHome: PropTypes.bool,
-    isDashboard: PropTypes.bool,
-    isLogin: PropTypes.bool
 };
 
 export default Navigation;
