+++ src/components/dashboard/bloodlines/Bloodlines.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import Navigation from './Navigation';
-import MessageContentContainer from '../containers/MessageContentContainer';
+import MessageContentContainer from '../../../containers/MessageContentContainer';
 
 
 class Bloodlines extends Component {
@@ -15,7 +14,6 @@ class Bloodlines extends Component {
     render() {
         return (
         	<div>
-        		<Navigation />
                 <MessageContentContainer url={this.state.url}/>
             </div>
         );
