+++ src/index.js
@@ -5,8 +5,7 @@ import About from './components/About';
 import Login from './components/Login';
 import Logout from './components/Logout';
 import Register from './components/Register';
-import Dashboard from './components/Dashboard';
-import Bloodlines from './components/Bloodlines'
+import Dashboard from './components/dashboard/Dashboard';
 
 import { Provider } from 'react-redux';
 import { Router, Route, browserHistory } from 'react-router';
@@ -18,7 +17,9 @@ import './index.css';
 const store = configureStore();
 
 function loggedIn() {
-    return false;
+    // TODO:  Fix auth
+
+    return true;
 }
 
 function requireAuth(nextState, replace) {
@@ -38,7 +39,6 @@ ReactDOM.render(
             <Route path="/logout" component={Logout} />
             <Route path="/register" component={Register} />
             <Route path="/dashboard" component={Dashboard} onEnter={requireAuth} />
-            <Route path="/bloodlines" component={Bloodlines}/>
         </Router>
     </Provider>,
     document.getElementById('root')
